Seperate sendToDIA, saveToJSON and getIntelligenceObject to utils
